split invoice detail
