+++ src/components/beta/my_account_page/create_team_modal.tsx
@@ -12,9 +12,16 @@ import PlanInfo from '@/components/beta/payment_page/plan_info';
 import PaymentInfo from '@/components/beta/payment_page/payment_info';
 import CreditCardInfo from '@/components/beta/payment_page/credit_card_info';
 import MessageModal from '@/components/message_modal/message_modal';
+import InvoiceDetail from '@/components/beta/invoice_page/invoice_detail';
 import { IMessageModal, MessageType } from '@/interfaces/message_modal';
+import {
+  IPlan,
+  IUserOwnedTeam,
+  ITeamInvoice,
+  TPaymentStatus,
+  TPlanType,
+} from '@/interfaces/subscription';
 import { APIName } from '@/constants/api_connection';
-import { IPlan, IUserOwnedTeam, TPaymentStatus, TPlanType } from '@/interfaces/subscription';
 import APIHandler from '@/lib/utils/api_handler';
 
 interface ICreateTeamModalProps {
@@ -93,7 +100,10 @@ const CreateTeamModal: React.FC<ICreateTeamModalProps> = ({ modalVisibilityHandl
   const [teamMemberInput, setTeamMemberInput] = useState<string>(''); // Info: (20250218 - Julian) input value
   const [teamMembers, setTeamMembers] = useState<string[]>([]);
   const [isValidEmail, setIsValidEmail] = useState<boolean>(true);
-  const [isHideArrow, setIsHideArrow] = useState<boolean>(false);
+  const [isHideArrow, setIsHideArrow] = useState<boolean>(false); // Info: (20250225 - Julian) 控制向下滾動的動畫
+  // ToDo: (20250225 - Julian) Implement API call
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const [teamInvoice, setTeamInvoice] = useState<ITeamInvoice | null>(null);
 
   // ToDo: (20250218 - Julian) For testing UI
   const fakeTeam = {
@@ -238,6 +248,8 @@ const CreateTeamModal: React.FC<ICreateTeamModalProps> = ({ modalVisibilityHandl
         ? teamMembers.length <= 0 // Info: (20250224 - Julian) 第二步 Member Email 必填
         : true;
 
+  const toTeamPage = () => window.open(`/team/${newTeam?.id}`, '_self');
+
   const toNextStep =
     currentStep === 1
       ? () => {
@@ -250,11 +262,11 @@ const CreateTeamModal: React.FC<ICreateTeamModalProps> = ({ modalVisibilityHandl
         }
       : currentStep === 2
         ? () => {
+            // Info: (20250225 - Julian) 第二步到第三步
             setCurrentStep(3);
           }
-        : // ToDo: (20250218 - Julian) Implement API call
-          // eslint-disable-next-line no-console
-          () => console.log('Create Team!');
+        : // Info: (20250225 - Julian) 第三步到 Team Page
+          toTeamPage;
 
   const cancelOrSkip =
     currentStep === 1
@@ -301,7 +313,7 @@ const CreateTeamModal: React.FC<ICreateTeamModalProps> = ({ modalVisibilityHandl
   });
 
   // Info: (20250224 - Julian) 訂閱方案
-  const subscriptionBody = newTeam && (
+  const subscriptionOverview = newTeam && (
     <div className="flex justify-center gap-lv-7">
       {PLANS.map((plan) => {
         const selectPlan = () => setSelectedPlan(plan);
@@ -315,12 +327,12 @@ const CreateTeamModal: React.FC<ICreateTeamModalProps> = ({ modalVisibilityHandl
             bordered
           />
         );
-      })}{' '}
+      })}
     </div>
   );
 
   // Info: (20250224 - Julian) 付款
-  const paymentBody = newTeam && selectedPlan && (
+  const paymentOverview = newTeam && selectedPlan && (
     <div className="flex min-h-600px w-900px gap-40px">
       <PlanInfo team={newTeam} plan={selectedPlan} />
 
@@ -361,6 +373,12 @@ const CreateTeamModal: React.FC<ICreateTeamModalProps> = ({ modalVisibilityHandl
   )} */}
     </div>
   );
+
+  const invoiceOverview = newTeam && teamInvoice && <InvoiceDetail invoice={teamInvoice} />;
+
+  const step3Body = selectedPlan
+    ? paymentOverview // Info: (20250225 - Julian) 有選擇方案 -> 顯示付款
+    : subscriptionOverview; // Info: (20250225 - Julian) 顯示訂閱方案
 
   const memberFormBody = (
     <div className="flex flex-col gap-8px text-sm">
@@ -410,10 +428,12 @@ const CreateTeamModal: React.FC<ICreateTeamModalProps> = ({ modalVisibilityHandl
     ) : currentStep === 2 ? (
       // Info: (20250217 - Julian) Member Email
       memberFormBody
+    ) : teamInvoice ? (
+      invoiceOverview // Info: (20250225 - Julian) 顯示 Invoice
     ) : (
       <>
-        {/* Info: (20250218 - Julian) Subscription Plan */}
-        {selectedPlan ? paymentBody : subscriptionBody}
+        {/* Info: (20250218 - Julian) Step 3 body */}
+        {step3Body}
 
         <ul className="ml-20px list-outside list-disc font-normal text-text-neutral-primary marker:text-surface-support-strong-maple">
           <li>{t('team:CREATE_TEAM_MODAL.PLAN_HINT')}</li>
@@ -465,16 +485,18 @@ const CreateTeamModal: React.FC<ICreateTeamModalProps> = ({ modalVisibilityHandl
 
           {/* Info: (20250217 - Julian) Buttons */}
           <div className="flex items-center justify-between">
-            {currentStep > 1 && (
+            {currentStep > 1 && !teamInvoice && (
               <Button type="button" variant="secondaryBorderless" onClick={backHandler}>
                 <TbArrowBackUp size={20} />
                 {t('common:COMMON.BACK')}
               </Button>
             )}
             <div className="ml-auto flex items-center gap-24px">
-              <Button type="button" variant="secondaryBorderless" onClick={cancelOrSkip}>
-                {cancelButtonText}
-              </Button>
+              {!teamInvoice && (
+                <Button type="button" variant="secondaryBorderless" onClick={cancelOrSkip}>
+                  {cancelButtonText}
+                </Button>
+              )}
               <Button
                 type="button"
                 variant="tertiary"
