+++ src/components/beta/my_account_page/team_item.tsx
@@ -3,14 +3,14 @@ import Image from 'next/image';
 import { useTranslation } from 'next-i18next';
 import { ITeam } from '@/interfaces/team';
 
-const TeamItem: React.FC<ITeam> = ({ name, profile, planType, role }) => {
+const TeamItem: React.FC<ITeam> = ({ name, imageId, planType, role }) => {
   const { t } = useTranslation(['team']);
 
   return (
     <div className="flex items-center gap-lv-7 rounded-sm border-2 border-stroke-neutral-quaternary bg-surface-neutral-surface-lv2 px-lv-5 py-lv-3">
       {/* Icon: (20250217 - Julian) Avatar */}
-      <div className="h-60px w-60px overflow-hidden rounded-sm border border-stroke-neutral-quaternary">
-        <Image src={profile.value} width={60} height={60} alt="team_avatar" />
+      <div className="h-60px w-60px place-content-center overflow-hidden rounded-sm border border-stroke-neutral-quaternary">
+        <Image src={imageId} width={60} height={60} alt="team_avatar" />
       </div>
       <div className="flex flex-1 items-center gap-lv-4">
         {/* Info: (20250217 - Julian) Team name */}
